PR commets addressed
