+++ src/index.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { BrowserRouter } from "react-router-dom";
 import "./index.css";
-import App from "./App";
-import reportWebVitals from "./reportWebVitals";
+import App from "./zz_TestRunner/App";
+import reportWebVitals from "./zz_TestRunner/reportWebVitals";
 
 ReactDOM.render(
   <React.StrictMode>
