+++ src/zz_TestRunner/Descriptions/CounterDescription/CounterDescription.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+const CounterDescription: React.FunctionComponent = () => {
+  return (
+    <div>
+      <p>
+        Open the Counter Component -<b>src/components/tests/Counter/Counter.tsx</b>
+      </p>
+      <p>It's is displayed on the right side.</p>
+    </div>
+  );
+};
+
+export default CounterDescription;
